// types/types.d.ts
export type GroupId = string | number;
export type GroupProjectId = string | number;
export type HookId = number;
export type ImpersonationTokenId = number;
export type IssueId = number;
export type JobId = number;
export type LabelId = number;
export type KeyId = string;
export type NamespaceId = string | number;
export type MergeRequestId = number;
export type MilestoneId = number;
export type NoteId = number;
export type PipelineId = number;
export type PipelineScheduleId = number;
export type ProjectId = string | number;
export type RunnerId = number;
export type SnippetId = number;
export type TodoId = number;
export type TriggerId = number;
export type VersionId = number;
export type UserId = number;

export interface Sudo {
  sudo?: string | number;
}

// Bundler
export interface Constructor {
  new (...args: any): any;
}

export type Mapper<T extends { [name: string]: Constructor }, P extends keyof T> = {
  [name in P]: InstanceType<T[name]>
};

export interface Bundle<T extends { [name: string]: Constructor }, P extends keyof T> {
  new (options?: any): Mapper<T, P>;
}

// Base Service
export interface BaseServiceOptions extends Sudo {
  oauthToken?: string;
  token?: string;
  jobToken?: string;
  host?: string;
  baseEndpoint?: string;
  version?: string;
  rejectUnauthorized?: boolean;
  requester?: {
    get: Function;
    post: Function;
    put: Function;
    delete: Function;
    stream: Function;
  };
}
// ...
export interface BaseRequestOptions extends Sudo {
  [key: string]: any;
}

export interface PaginatedRequestOptions extends BaseRequestOptions {
  showPagination?: boolean;
  maxPages?: number;
  page?: number;
  perPage?: number;
}
// ...
export type EventAction =
  | 'created'
  | 'updated'
  | 'closed'
  | 'reopened'
  | 'pushed'
  | 'commented'
  | 'merged'
  | 'joined'
  | 'left'
  | 'destroyed'
  | 'expired';

export type EventTarget =
  | 'issue'
  | 'milestone'
  | 'merge_request'
  | 'note'
  | 'project'
  | 'snippet'
  | 'user';

export interface EventOptions {
  action?: EventAction;
  targetType?: EventTarget;
}
// ...
export interface ProjectUploadMetadata {
  filename?: string,
  contentType?: string,
}